backend/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
deepface
opencv-python
pandas
scikit-learn
tensorflow
